create separate module for CERTABO board specific code
